Moving validations to separate module, more WIP refactoring
