+++ lib/smart_proxy_ipam/phpipam/phpipam_client.rb
@@ -8,6 +8,7 @@ require 'uri'
 require 'sinatra'
 require 'smart_proxy_ipam/ipam'
 require 'smart_proxy_ipam/ipam_helper'
+require 'smart_proxy_ipam/ipam_validator'
 require 'smart_proxy_ipam/api_resource'
 require 'smart_proxy_ipam/ip_cache'
 
@@ -16,23 +17,24 @@ module Proxy::Phpipam
   class PhpipamClient
     include Proxy::Log
     include Proxy::Ipam::IpamHelper
+    include Proxy::Ipam::IpamValidator
 
     @ip_cache = nil
-    MAX_RETRIES = 5
 
     def initialize(conf)
       @conf = conf
       @api_base = "#{@conf[:url]}/api/#{@conf[:user]}/"
-      @token = authenticate('/user/')
+      @token = authenticate
       @api_resource = Proxy::Ipam::ApiResource.new(api_base: @api_base, token: @token, auth_header: 'Token')
       @ip_cache = Proxy::Ipam::IpCache.new(provider: 'phpipam')
     end
 
-    def get_ipam_subnet(cidr, group_id = nil)
-      if group_id.nil? || group_id.empty?
+    def get_ipam_subnet(cidr, group_name = nil)
+      if group_name.nil?
         get_ipam_subnet_by_cidr(cidr)
       else
-        get_ipam_subnet_by_group(cidr, group_id)
+        group = get_ipam_group(group_name)
+        get_ipam_subnet_by_group(cidr, group[:id])
       end
     end
 
@@ -79,7 +81,7 @@ module Proxy::Phpipam
       return nil if group_name.nil?
       group = @api_resource.get("sections/#{group_name}/")
       json_body = JSON.parse(group.body)
-      return nil if json_body['data'].nil?
+      raise errors[:no_group] if json_body['data'].nil?
 
       data = {
         id: json_body['data']['id'],
@@ -158,7 +160,7 @@ module Proxy::Phpipam
       subnet_hash = @ip_cache.get_cidr(group, cidr)
       next_ip = nil
 
-      return { message: json_body['message'] } if json_body['message']
+      return { error: json_body['message'] } if json_body['message']
 
       if subnet_hash&.key?(mac.to_sym)
         next_ip = @ip_cache.get_ip(group, cidr, mac)
@@ -170,11 +172,12 @@ module Proxy::Phpipam
           next_ip = new_ip.to_s
           @ip_cache.add(new_ip, mac, cidr, group)
         else
-          next_ip = find_new_ip(subnet[:id], new_ip, mac, cidr, group)
+          next_ip = find_new_ip(@ip_cache, subnet[:id], new_ip, mac, cidr, group)
         end
 
-        return { error: "Unable to find another available IP address in subnet #{cidr}" } if next_ip.nil?
-        return { error: "It is possible that there are no more free addresses in subnet #{cidr}. Available IP's may be cached, and could become available after in-memory IP cache is cleared(up to #{@ip_cache.get_cleanup_interval} seconds)." } unless usable_ip(next_ip, cidr)
+        unless usable_ip(next_ip, cidr)
+          return { error: "No free addresses found in subnet #{cidr}. Some available ip's may be cached. Try again in #{@ip_cache.get_cleanup_interval} seconds after cache is cleared." }
+        end
       end
 
       return nil if no_free_ip_found?(next_ip)
@@ -186,6 +189,10 @@ module Proxy::Phpipam
       ip.is_a?(Hash) && ip['message'] && ip['message'].downcase == 'no free addresses found'
     end
 
+    def subnet_exists?(subnet)
+      !(subnet[:message] && subnet[:message].downcase == 'no subnet found')
+    end
+
     def groups_supported?
       true
     end
@@ -194,12 +201,10 @@ module Proxy::Phpipam
       !@token.nil?
     end
 
-    def subnet_exists?(subnet)
-      !(subnet[:message] && subnet[:message].downcase == 'no subnet found')
-    end
+    private
 
-    def authenticate(path)
-      auth_uri = URI(@api_base + path)
+    def authenticate
+      auth_uri = URI(@api_base + '/user/')
       request = Net::HTTP::Post.new(auth_uri)
       request.basic_auth @conf[:user], @conf[:password]
 
@@ -210,36 +215,6 @@ module Proxy::Phpipam
       response = JSON.parse(response.body)
       logger.warn(response['message']) if response['message']
       response.dig('data', 'token')
-    end
-
-    # Called when next available IP from external IPAM has been cached by another user/host, but
-    # not actually persisted in external IPAM. Will increment the IP(MAX_RETRIES times), and
-    # see if it is available in external IPAM.
-    def find_new_ip(subnet_id, ip, mac, cidr, group_name)
-      found_ip = nil
-      temp_ip = ip
-      retry_count = 0
-
-      loop do
-        new_ip = increment_ip(temp_ip)
-        ipam_ip = ip_exists?(new_ip, subnet_id, group_name)
-
-        # If new IP doesn't exist in IPAM and not in the cache
-        if !ipam_ip && !@ip_cache.ip_exists(new_ip, cidr, group_name)
-          found_ip = new_ip.to_s
-          @ip_cache.add(found_ip, mac, cidr, group_name)
-          break
-        end
-
-        temp_ip = new_ip
-        retry_count += 1
-        break if retry_count >= MAX_RETRIES
-      end
-
-      # Return the original IP found in external ipam if no new ones found after MAX_RETRIES
-      return ip if found_ip.nil?
-
-      found_ip
     end
   end
 end
